Sources/requirements.txt
Flask==1.1.2
Flask_RESTful==0.3.8
gevent==20.9.0
numpy==1.18.1
elasticsearch==7.9.1
annoy==1.16.3
opencv_python==4.3.0.36
pyzmq==20.0.0
